### 2016-08-16 #50 @arfon
@n3mo thanks for this submission! Before we can proceed would you mind extracting the references into a `paper.bib` file. We won't be able to process this submission until we can compile the paper using pandoc.
