+++ paper/paper.md
@@ -1,0 +1,25 @@
+---
+title: 'MassMine: Your Access To Data'
+tags:
+  - social media
+  - mining
+  - api
+  - data
+  - web scraping
+authors:
+ - name: Nicholas M Van Horn
+   orcid: 0000-0002-0695-7765
+   affiliation: Capital University
+ - name: Aaron Beveridge
+   orcid: 0000-0002-5156-3044
+   affiliation: University of Florida
+date: 15 August 2016
+bibliography: paper.bib
+---
+
+# Summary
+
+MassMine [@VanHornBeveridge2016] is a research tool designed to simplify the collection and use of data from social media and news websites, such as Twitter, Tumblr, Google Trends, Wikipedia, as well as web URL scraping. MassMine runs on personal computers and in "headless" environments (i.e., servers/clusters). In addition to searching, downloading, parsing, and storing data, MassMine handles any authentication credentials on behalf of the user, allowing users to focus on their research rather than programming code to negotiate with online APIs. Further, MassMine automatically monitors and respects rate limits commonly imposed by online services, ensuring that a data collection project does not violate usage rights set in terms of service agreements. Custom data requests are made simple, limited only by the restrictions of the data providers, allowing for flexible deployment for individualized research needs. All data collected is managed automatically for the user. This means that all tweets, trends, posts, etc. related to users' research questions are saved permanently on local hard drives. Built to accommodate short and long-lived data collection projects, MassMine makes it easy to create a database of social media content spanning days, months, or years. Visit the official online website [@MassMine2016] for downloads, as well as installation and usage instructions.
+
+# References
+
