+++ paper/paper.bib
@@ -1,0 +1,16 @@
+@online{VanHornBeveridge2016,
+  title =	 {MassMine: Social Media and Online Web Content Mining
+                  Software},
+  author =	 {Van Horn, Nicholas M. and Beveridge, Aaron},
+  url =		 {https://github.com/n3mo/massmine},
+  urldate =	 {2016-08-15},
+  year =	 {2016}
+}
+
+@online{MassMine2016,
+  title =	 {MassMine Official Website},
+  author =	 {Van Horn, Nicholas M. and Beveridge, Aaron},
+  url =		 {http://www.massmine.org},
+  urldate =	 {2016-08-15}
+  year =	 {2016}
+}
