### 2016-08-27 2d79625
Changelog updated
