### 2016-08-29 5f34d25
ssl shutdown timeout added to changelog
