### 2016-09-08 cb860aa
Release ver number added to changelog
