### 2016-09-19 #50 @labarba
You can do both. For example, I'm lead author of this paper (now accepted) and I [pinged](https://twitter.com/LorenaABarba/status/756423713466318848) someone on Twitter asking for a second reviewer:
https://github.com/openjournals/joss-reviews/issues/43
