### 2016-11-04 #50 @julianmcauley
Comments on the Checklist:

**Installation: Does installation proceed as outlined in the documentation?**
Installation instructions aren't really provided in the release, but rather on a link in the readme. Following that link leads to clear instructions.

**Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.**
The dependencies are clearly stated on the webpage linked from the readme.

**Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).**
Somewhat. Examples are given in the form of youtube videos containing walked-through examples. Personally I find this a very difficult to navigate means of introducing people to the software. I would like to see such videos *in addition to* (rather than instead of) actual code examples.

**Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g. API method documentation)?**
Documentation is okay. Mostly it's in the form of comments within the .scm code. There's not really organized documentation in a single document (that I can find?) though. Having no experience with scheme I'll admit that I find this a little difficult to parse.

**Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?**
No (not that I can find). Automated tests here would be useful -- some of the functionality being implemented here depends on APIs from external websites that may change from time to time, causing the software to break. As it is I'm not sure how users will figure out that this has happened (other than the software failing).

**Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**
No, I was not able to find this documentation.

**References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?**
N/A

General comments:
Crawling large datasets is a cumbersome and time consuming task, and one that it's valuable to build tools to automate. This software focuses on a few specific websites (google, tumbler, twitter, wikipedia), though potentially more could be added following the same framework in the future. The software seems easy enough to use, and the documentation is okay, but depends heavily on reading code and watching videos rather than (say) providing code examples directly which might be easier.

Positively, I think this is a well put-together project that contains features that people would be interested in. Negatively, the focus so far is on websites that already have strong API support, so really this code is adding an additional layer on top of an easy-to-use API. To somebody like me (who has never before read scheme code), it would be easier to follow the documentation from these websites' APIs directly rather than following this code. But maybe I'm in the minority. Certainly this wouldn't be an issue once more "hard to crawl" websites are added into the mix, if that's the plan.
