### 2016-11-04 fd080f3
Contribution and support info added
