### 2016-11-08 #50 @n3mo
To my knowledge they do not offer test credentials.
