### 2016-11-20 76b0553
Unit tests added to changelog
