### 2016-11-20 #50 @n3mo
Tests are now available! They are included in a separate directory in the repo, but can be ran directly with massmine. The [installation instructions](http://www.massmine.org/docs/install.html) have been updated with details. But running the tests is simple once all build dependencies are installed:

```sh
./massmine.scm --test ./tests/run.scm
```
