### 2016-11-21 #50 @mgymrek
Great! @julianmcauley would you be able to take a look at the added tests?
