### 2016-11-29 #50 @mgymrek
@n3mo I have also taken a look at the tests and am not totally sure what is being tested. Could you provide a brief description here?
