### 2016-12-01 #50 @mgymrek
Thanks @n3mo for the explanation. This sounds reasonable to me. @arfon, I think we can accept this. what is the next step?
