### 2016-12-06 #50 @n3mo
I've archived the software with Zenodo. The DOI is: 10.5281/zenodo.193078
